Bind mixed-type GPU arenas once and validate only executed schedules
Reinterpret typed views through one physical storage binding, preserve packed half writes atomically, and keep cooperative staging disjoint per subgroup. Add quantized gradient regressions for f32 and f16 and include them in browser checks.

Exclude abandoned search candidates from schedule validation while retaining them for autotuning. Fix strict lint/format checks and give software-GPU training tests an appropriate timeout and concurrency limit.

+++ fusor/fusor/src/program.rs
@@ -47,7 +47,7 @@ impl TrainingProgram {
         let limits = target.device().device().limits();
         let max_bytes = limits
             .max_buffer_size
-            .min(u64::from(limits.max_storage_buffer_binding_size));
+            .min(limits.max_storage_buffer_binding_size);
         let ids = roots.iter().map(Dyn::id).collect::<Vec<_>>();
         let state = feedback
             .iter()
